pyproject.toml
[tool.poetry]
name = "rate"
version = "0.1.0"
description = ""
authors = ["sanch13 <sanch13sanch@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
loguru = "^0.6.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
